+++ src/solo/mod.rs
@@ -13,6 +13,7 @@ pub mod cell_reads;
 pub mod count;
 pub mod gene;
 pub mod libcxx_rng;
+pub mod sgt;
 pub mod smartseq;
 pub mod whitelist;